requirements.txt
--prefer-binary
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
tzdata==2023.3
Werkzeug==2.3.4
